% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{CpGsDemo}
\alias{CpGsDemo}
\title{Toy data for the package - 1000 CpG regions to be tested in a GRanges object}
\description{
Toy data for the package - 1000 CpG regions to be tested in a GRanges object
}
\author{
Tom Mayo
}
\keyword{data}

