% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/determineGroupComps.R
\name{determineGroupComps}
\alias{determineGroupComps}
\title{Creates strings of sample pair comparisons}
\usage{
determineGroupComps(samples1, samples2 = NULL, type)
}
\arguments{
\item{samples1}{A vector of sample names from one group}

\item{samples2}{A vector of sample names from the other group, if we want to
specify between-group comparisons}

\item{type}{'within' or 'between'. 'within' returns all the sample pairs
within samples1, 'between' returns all the sample pairs
between samples1 and samples2}
}
\value{
A vector of sample pair comparisons of the form 'sample1 vs sample2'
 for use with the M3D functions
}
\description{
Takes in a vector of strings of sample names and returns strings of
all the comparisons, either within a testing group or between testing groups.
 This is not intended to be called directly by the user.
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

