% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComps.R
\name{findComps}
\alias{findComps}
\title{Finds columns in the M3D test-statistic matrix}
\usage{
findComps(MMD, samples)
}
\arguments{
\item{MMD}{A matrix containing the M3D test-statistic, the difference the
 full and methylation blind metrics,
for each region in the CpGs object. Each column is a comparison between two
 samples, which are described in the column names.}

\item{samples}{A vector of sample pairs of the form 'sample1 vs sample2' as
returned from determineGroupComps}
}
\value{
Returns the indices of the M3D test-statistic components that contain
 the sample pair comparisons in 'samples'
}
\description{
Returns the columns of the test-statistic matrix that refer to specific
 samples. This is not intended to be called directly by the user.
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

