% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3D_Para_lite.R
\name{M3D_Para_lite}
\alias{M3D_Para_lite}
\title{Computes the components of the M3D test-statistic over all regions for all
 sample-pairs.}
\usage{
M3D_Para_lite(rrbs, CpGs, overlaps, group1 = NaN, group2 = NaN,
  num.cores = NaN)
}
\arguments{
\item{rrbs}{An rrbs object containing methylation and coverage data as
created using the BiSeq pacakge}

\item{CpGs}{A GRanges object detailing the testing regions.}

\item{overlaps}{The overlaps between the list of testing regions and the
 methylation data. This is obtained using the
function findOverlaps(CpGs,rrbs) for a GRanges object CpGs detailing the
 testing regions.}

\item{group1}{The name of the first group for the comparison. This is stored
in colData(rrbs). Default finds first unique group in colData(rrbs).}

\item{group2}{The name of the second group for the comparison. This is stored
in colData(rrbs). Default finds second unique group in colData(rrbs).}

\item{num.cores}{Integer giving the number or cores to use. Defaults to the
maximum available}
}
\value{
This returns the two components of the M3D test-statistic for each
 region over all sample pairs as a matrix.
Subtracting them gives the M3D test-statistic. This is processed with the
 function pvals.
}
\description{
Parallel implementation of M3D_Wrapper_lite function.
Returns the M3D test-statistic, without the compenents, for all regions
with the cross group sample pairs averaged to save memory (in column one),
as a matrix.
}
\examples{
data(rrbsDemo)
data(CpGsDemo)
CpGsDemo <- CpGsDemo[1:20]
overlaps <- GenomicRanges::findOverlaps(CpGsDemo,rrbsDemo)
M3D_list <- M3D_Para_lite(rrbsDemo,CpGsDemo, overlaps)
head(M3D_list)
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}
\references{
Gretton, A., Borgwardt, K. M., Rasch, M., Scholkopf, B., Smola,
 A. J. (2006). A kernel method for the two-sample-problem. In Advances in
  neural information processing systems (pp. 513-520).
}

