% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/M3D_Single.R
\name{M3D_Single}
\alias{M3D_Single}
\title{Computes the components of the M3D test-statistic over one region for 2
 samples}
\usage{
M3D_Single(testData, locMx, locInds, method = "MinusCovMMD")
}
\arguments{
\item{testData}{Contains the methylation data over two samples for a given
region.}

\item{locMx}{The matrix of distances between the CpG sites.}

\item{locInds}{The indices of the non-zero entries of locMx.}

\item{method}{This specifies whether to return the full MMD and the
 methylation blind MMD (focusing only on the coverage)
or just the former. if method ='MinusCovMMD' it is both, all other values
 return just the full MMD.}
}
\value{
This returns the value of the MMD for the region between the two
 samples as a numeric. If method
is set to 'MinusCovMMD', a list is returned of the full MMD and the coverage
 only MMD. Subtracting them gives the M3D test-statistic.
}
\description{
Returns the two components of the M3D test-statistic - the MMD
 (Gretton et al. 2006) for the full data and the coverge
only data, respectively. This is not intended to be called directly
 by the user.
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}
\references{
Gretton, A., Borgwardt, K. M., Rasch, M., Scholkopf, B., Smola,
 A. J. (2006). A kernel method for the two-sample-problem. In Advances in
  neural information processing systems (pp. 513-520).
}

