% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/M3D_Wrapper.R
\name{M3D_Wrapper}
\alias{M3D_Wrapper}
\title{Computes the components of the M3D test-statistic over all regions for all
 sample-pairs.}
\usage{
M3D_Wrapper(rrbs, overlaps, para = FALSE)
}
\arguments{
\item{rrbs}{An rrbs object containing methylation and coverage data as
created using the BiSeq pacakge}

\item{overlaps}{The overlaps between the list of testing regions and the
 methylation data. This is obtained using the
function findOverlaps(CpGs,rrbs) for a GRanges object CpGs detailing the
 testing regions.}

\item{para}{Set to true if called via M3D_Para}
}
\value{
This returns the two components of the M3D test-statistic for each
 region over all sample pairs as a matrix.
Subtracting them gives the M3D test-statistic. This is processed with the
 function pvals.
}
\description{
Returns the two components of the M3D test-statistic - the MMD
 (Gretton et al. 2006) for the full data and the coverge
only data, respectively - for all regions and all samples pairs, as a matrix.
}
\examples{
\donttest{data(rrbsDemo)
data(CpGsDemo)
overlaps <- findOverlaps(CpGsDemo,rrbsDemo)
M3D_list <- M3D_Wrapper(rrbsDemo,overlaps)
head(M3d_list$Full-M3D_list$Coverage)}
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}
\references{
Gretton, A., Borgwardt, K. M., Rasch, M., Scholkopf, B., Smola,
 A. J. (2006). A kernel method for the two-sample-problem. In Advances in
  neural information processing systems (pp. 513-520).
}

