% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{median_freq}
\alias{median_freq}
\title{Finds the median}
\usage{
median_freq(values, freqs)
}
\arguments{
\item{values}{A vector of the unique values that occur}

\item{freqs}{A vector of the number of occurrences of each value}
}
\value{
Returns the median value of the data comprising each entry in values
 repeated the corresponding entry in freqs number of times, as a numeric.
}
\description{
Returns the median of a list of values with corresponding frequencies. This
 is not intended to be called directly by the user.
}
\examples{
median_freq(c(1,2,3), c(3,6,10))
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

