% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/medianFreq.R
\name{medianFreq}
\alias{medianFreq}
\title{Finds the median}
\usage{
medianFreq(values, freqs)
}
\arguments{
\item{values}{A vector of the unique values that occur}

\item{freqs}{A vector of the number of occurrences of each value}
}
\value{
Returns the median value of the data comprising each entry in values
 repeated the corresponding entry in freqs number of times, as a numeric.
}
\description{
Returns the median of a list of values with corresponding frequencies. This
 is not intended to be called directly by the user.
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

