% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{MMDlistDemo}
\alias{MMDlistDemo}
\title{Toy data for the package - the output of the M3D_Wrapper function.}
\description{
Toy data for the package - the output of the M3D_Wrapper function.
}
\author{
Tom Mayo
}
\keyword{data}

