% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComps.R
\docType{data}
\name{PDemo}
\alias{PDemo}
\title{Toy data for the package - the output of the pvals function.}
\description{
Toy data for the package - the output of the pvals function.
}
\author{
Tom Mayo
}
\keyword{data}

