% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMethProfile.R
\name{plotMethProfile}
\alias{plotMethProfile}
\title{Plots methylation profiles over a specific region}
\usage{
plotMethProfile(rrbs, CpGs, group1, group2, CpGindex)
}
\arguments{
\item{rrbs}{An rrbs object containing methylation and coverage data as
created using the BiSeq pacakge}

\item{CpGs}{A GRanges object with each row being a testing region}

\item{group1}{The name of the first testing group}

\item{group2}{The name of the second testing group}

\item{CpGindex}{The index within the CpGs object of the region we are
plotting}
}
\value{
NULL, the function plots the profiles
}
\description{
Plots a smoothed methylation profile for each of the two testing groups.
Within each group, the mean of
methylation level is taken, smoothed and plotted, along with the individual
values.
}
\examples{
# plot the 9th region in the Toy Data Set
data(rrbsDemo)
data(CpGsDemo)
plotMethProfile(rrbsDemo, CpGsDemo, 'H1-hESC', 'K562', 9)
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

