% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pvals}
\alias{pvals}
\title{Computes p-values}
\usage{
pvals(rrbs, CpGs, MMD, group1, group2, smaller = FALSE,
  comparison = "allReps")
}
\arguments{
\item{rrbs}{An rrbs object containing methylation and coverage data as
created using the BiSeq pacakge}

\item{CpGs}{A GRanges object with each row being a testing region}

\item{MMD}{A matrix containing the M3D test-statistic, the difference the
 full and methylation blind metrics,
for each region in the CpGs object. Each column is a comparison between two
 samples, which are described in the column names.}

\item{group1}{The name of the first group for the comparison. This is stored
in colData(rrbs)}

\item{group2}{The name of the second group for the comparison. This is stored
in colData(rrbs)}

\item{smaller}{Determines whether the p-value is computed whether the
 test-statistic is greater or lesser than inter-replicate
values. For our purposes, it should be set to FALSE.}

\item{comparison}{Details which groups we are using to define our empirical
 testing distribution. The default is to
use all of them, however, should the user find one group contains unusually
 high variability, then that group can be selected.
Values are either 'allReps', 'Group1' or 'Group2'.}
}
\value{
Returns a list P, with 2 entries. 'FDRmean' is the Benjamini-Hochberg
 adjusted p-values. The unadjusted p-values are stored in 'Pmean'.
}
\description{
Returns p-values for each region reflecting the probability of observing the
 mean test-statistic
of the between group comparisons among the inter-replicate comparisons.
}
\examples{
\donttest{data(rrbsDemo)
data(CpGsDemo)
data(MMDlistDemo)
M3Dstat <- MMDlistDemo$Full-MMDlistDemo$Coverage
group1 <- unique(colData(rrbsDemo)$group)[1]
group2 <-unique(colData(rrbsDemo)$group)[2]
PDemo <- pvals(rrbsDemo, CpGsDemo, M3Dstat,
            group1, group2, smaller=FALSE,comparison='allReps')}
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

