% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals_lite.R
\name{pvals_lite}
\alias{pvals_lite}
\title{Computes p-values}
\usage{
pvals_lite(rrbs, CpGs, M3D_stat_lite, group1, group2, smaller = FALSE,
  comparison = "allReps", method = "empirical", closePara = 0.005)
}
\arguments{
\item{rrbs}{An rrbs object containing methylation and coverage data as
created using the BiSeq pacakge}

\item{CpGs}{A GRanges object with each row being a testing region}

\item{M3D_stat_lite}{A matrix containing the M3D test-statistic, as returned
from the M3D_Wrapper_lite function. This has the mean of the MMDs for the
comparisons between groups in the first column, and the within groups MMDs
in the others, for each region in the CpGs object.}

\item{group1}{The name of the first group for the comparison. This is stored
in colData(rrbs)}

\item{group2}{The name of the second group for the comparison. This is stored
in colData(rrbs)}

\item{smaller}{Determines whether the p-value is computed whether the
 test-statistic is greater or lesser than inter-replicate
values. For our purposes, it should be set to FALSE.}

\item{comparison}{Details which groups we are using to define our empirical
 testing distribution. The default is to
use all of them, however, should the user find one group contains unusually
 high variability, then that group can be selected.
Values are either 'allReps', 'Group1' or 'Group2'.}

\item{method}{Determines which method is used to calculate p-values.
'empirical' uses the empirical distribution directly, without modelling.
 This is the default. 'model', fits an exponential distribution to the tail
 of our null distribution.}

\item{closePara}{Sets a threshold for how close the exponential curve should
fit the empirical distribution in the 'model' method. If the method produces
errors, consider raising this parameter.}
}
\value{
Returns a list P, with 2 entries. 'FDRmean' is the Benjamini-Hochberg
 adjusted p-values. The unadjusted p-values are stored in 'Pmean'.
}
\description{
Returns p-values for each region reflecting the probability of observing the
 mean test-statistic
of the between group comparisons among the inter-replicate comparisons. The
'lite' version works with the outcome of M3D_Wrapper_lite and the 'lite'
parallel version.
}
\examples{
data(rrbsDemo)
data(CpGsDemo)
CpGsDemo <- CpGsDemo[1:5]
overlaps <- GenomicRanges::findOverlaps(CpGsDemo,rrbsDemo)
M3D_list <- M3D_Wrapper_lite(rrbsDemo,overlaps)
group1 <- unique(GenomicRanges::colData(rrbsDemo)$group)[1]
group2 <-unique(GenomicRanges::colData(rrbsDemo)$group)[2]
PDemo <- pvals_lite(rrbsDemo, CpGsDemo, M3D_list,
            group1, group2, smaller=FALSE,comparison='allReps')
head(PDemo)
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

