% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readENCODEdata}
\alias{readENCODEdata}
\title{Reads in ENCODE RRBS data}
\usage{
readENCODEdata(files, colData)
}
\arguments{
\item{files}{A character pointing the the rrbs files downloads from the
ENCODE database.}

\item{colData}{Samples' names plus additional sample information as
character, data.frame or DataFrame.}
}
\value{
Returns a BSraw object storing methylation and coverage data -
 the underlying structure for this package.
}
\description{
Reads in RRBS data in bed file format from the ENCODE consortium and outputs
 an rrbs data structure. Adapted from readBismark in the BiSeq package.
}
\examples{
\donttest{
# download the files and change the working directory
# to that location
files <- c('wgEncodeHaibMethylRrbsH1hescHaibSitesRep1.bed.gz',
'wgEncodeHaibMethylRrbsH1hescHaibSitesRep2.bed.gz',
'wgEncodeHaibMethylRrbsK562HaibSitesRep1.bed.gz',
'wgEncodeHaibMethylRrbsK562HaibSitesRep2.bed.gz')
group <- factor(c('H1-hESC','H1-hESC','K562','K562'))
samples <- c('H1-hESC1','H1-hESC2','K562-1','K562-2')
colData <- DataFrame(group,row.names= samples)
rrbs <- readENCODEdata(files,colData)}
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

