% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{rrbsDemo}
\alias{rrbsDemo}
\title{Toy data for the package - methylation data for cytosines sites within the testing regions only, in an rrbs object.}
\description{
Toy data for the package - methylation data for cytosines sites within the testing regions only, in an rrbs object.
}
\author{
Tom Mayo
}
\keyword{data}

