% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_colData_structure.R
\name{test_colData_structure}
\alias{test_colData_structure}
\title{Tests that the files are specified in a workable order for M3D to run}
\usage{
test_colData_structure(colData)
}
\arguments{
\item{colData}{Samples' names plus additional sample information as
character, data.frame or DataFrame.}
}
\value{
Returns TRUE if the colData is structured correctly for downstream
analysis, FALSE otherwise
}
\description{
M3D requires that the files be specified group by group (all of condition 1,
then all of condition 2 etc.). This function tests this, and is used before
loading so that the error is caught before time is wasted.
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

