% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/M3D_Para.R
\name{M3D_Para}
\alias{M3D_Para}
\title{Computes the components of the M3D test-statistic over all regions for all
 sample-pairs.}
\usage{
M3D_Para(rrbs, CpGs, overlaps, num.cores = NaN)
}
\arguments{
\item{rrbs}{An rrbs object containing methylation and coverage data as
created using the BiSeq pacakge}

\item{CpGs}{A GRanges object detailing the testing regions.}

\item{overlaps}{The overlaps between the list of testing regions and the
 methylation data. This is obtained using the
function findOverlaps(CpGs,rrbs) for a GRanges object CpGs detailing the
 testing regions.}

\item{num.cores}{Integer giving the number or cores to use. Defaults to the
maximum available}
}
\value{
This returns the two components of the M3D test-statistic for each
 region over all sample pairs as a matrix.
Subtracting them gives the M3D test-statistic. This is processed with the
 function pvals.
}
\description{
Parallel implementation of M3D_Wrapper function.
Returns the two components of the M3D test-statistic - the MMD
 (Gretton et al. 2006) for the full data and the coverge
only data, respectively - for all regions and all samples pairs, as a matrix.
}
\examples{
\donttest{data(rrbsDemo)
data(CpGsDemo)
M3D_list <- M3D_Para(rrbsDemo,CpGsDemo)
head(M3d_list$Full-M3D_list$Coverage)}
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}
\references{
Gretton, A., Borgwardt, K. M., Rasch, M., Scholkopf, B., Smola,
 A. J. (2006). A kernel method for the two-sample-problem. In Advances in
  neural information processing systems (pp. 513-520).
}

