% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComps.R
\docType{data}
\name{MMDlistDemo}
\alias{MMDlistDemo}
\title{Toy data for the package - the output of the M3D_Wrapper function.}
\description{
Toy data for the package - the output of the M3D_Wrapper function.
}
\author{
Tom Mayo
}
\keyword{data}

