% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findComps.R
\docType{data}
\name{rrbsDemo}
\alias{rrbsDemo}
\title{Toy data for the package - methylation data for cytosines sites within the testing regions only, in an rrbs object.}
\description{
Toy data for the package - methylation data for cytosines sites within the testing regions only, in an rrbs object.
}
\author{
Tom Mayo
}
\keyword{data}

